Push discovery oncreate and get current logic into Helper object to make testing more stright forward. Dont remove keys before initial connection has been established.
