+++ tessera-jaxrs/sync-jaxrs/src/test/java/com/quorum/tessera/p2p/PartyInfoBroadcasterTest.java
@@ -170,8 +170,8 @@ public class PartyInfoBroadcasterTest {
 
         partyInfoBroadcaster.pollSingleParty(uriData,"somebytes".getBytes());
 
-        verify(discovery).onDisconnect(URI.create(uriData));
         verify(p2pClient).sendPartyInfo(anyString(),any(byte[].class));
+        assertThat(connectedNodes).isEmpty();
 
     }
 }
