--- tessera-partyinfo/src/main/java/com/quorum/tessera/discovery/OnCreateHelper.java
@@ -1,12 +1,0 @@
-package com.quorum.tessera.discovery;
-
-import java.util.ServiceLoader;
-
-public interface OnCreateHelper {
-
-    void onCreate();
-
-    static OnCreateHelper getInstance() {
-        return ServiceLoader.load(OnCreateHelper.class).findFirst().get();
-    }
-}
