+++ tessera-partyinfo/src/main/resources/META-INF/services/com.quorum.tessera.discovery.DiscoveryHelper
@@ -1,0 +1,1 @@
+com.quorum.tessera.discovery.DiscoveryHelperFactory
