--- tessera-partyinfo/src/test/java/com/quorum/tessera/discovery/MockOnCreateHelper.java
@@ -1,21 +1,0 @@
-package com.quorum.tessera.discovery;
-
-import java.util.concurrent.atomic.AtomicInteger;
-
-public class MockOnCreateHelper implements OnCreateHelper {
-
-    private static final AtomicInteger COUNTER = new AtomicInteger(0);
-
-    @Override
-    public void onCreate() {
-        COUNTER.incrementAndGet();
-    }
-
-    static void reset() {
-        COUNTER.set(0);
-    }
-
-    int getInvocationCount() {
-        return COUNTER.get();
-    }
-}
