--- tessera-partyinfo/src/test/resources/META-INF/services/com.quorum.tessera.discovery.OnCreateHelper
@@ -1,1 +1,0 @@
-com.quorum.tessera.discovery.MockOnCreateHelper
